Move TransactionOutbox to interface
