+++ pom.xml
@@ -37,7 +37,7 @@
     <logback.version>1.2.3</logback.version>
     <revision>1.3.99999-SNAPSHOT</revision>
     <junit.jupiter.version>5.7.0</junit.jupiter.version>
-    <testcontainers.version>1.14.3</testcontainers.version>
+    <testcontainers.version>1.15.0</testcontainers.version>
   </properties>
 
   <modules>
