+++ transactionoutbox-core/src/main/java/com/gruelbox/transactionoutbox/TransactionOutbox.java
@@ -1,26 +1,10 @@
 package com.gruelbox.transactionoutbox;
 
-import static com.gruelbox.transactionoutbox.Utils.logAtLevel;
-import static com.gruelbox.transactionoutbox.Utils.uncheckedly;
-import static java.time.temporal.ChronoUnit.MILLIS;
-import static java.time.temporal.ChronoUnit.MINUTES;
-
-import java.lang.reflect.InvocationTargetException;
 import java.time.Clock;
 import java.time.Duration;
-import java.time.Instant;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.UUID;
 import java.util.concurrent.Executor;
 import javax.validation.ClockProvider;
-import javax.validation.Valid;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotNull;
 import lombok.ToString;
-import lombok.extern.slf4j.Slf4j;
-import org.hibernate.validator.constraints.Length;
-import org.hibernate.validator.internal.engine.DefaultClockProvider;
 import org.slf4j.MDC;
 import org.slf4j.event.Level;
 
@@ -30,63 +14,11 @@ import org.slf4j.event.Level;
  * pattern for Java. See <a href="https://github.com/gruelbox/transaction-outbox">README</a> for
  * usage instructions.
  */
-@Slf4j
-public class TransactionOutbox {
-
-  private static final int DEFAULT_FLUSH_BATCH_SIZE = 4096;
-
-  @NotNull private final TransactionManager transactionManager;
-  @Valid @NotNull private final Persistor persistor;
-  @Valid @NotNull private final Instantiator instantiator;
-  @NotNull private final Submitter submitter;
-  @NotNull private final Duration attemptFrequency;
-  @NotNull private final Level logLevelTemporaryFailure;
-
-  @Min(1)
-  private final int blockAfterAttempts;
-
-  @Min(1)
-  private final int flushBatchSize;
-
-  @NotNull private final ClockProvider clockProvider;
-  @NotNull private final TransactionOutboxListener listener;
-  private final boolean serializeMdc;
-  private final Validator validator;
-  @NotNull private final Duration retentionThreshold;
-
-  private TransactionOutbox(
-      TransactionManager transactionManager,
-      Instantiator instantiator,
-      Submitter submitter,
-      Duration attemptFrequency,
-      int blockAfterAttempts,
-      int flushBatchSize,
-      ClockProvider clockProvider,
-      TransactionOutboxListener listener,
-      Persistor persistor,
-      Level logLevelTemporaryFailure,
-      Boolean serializeMdc,
-      Duration retentionThreshold) {
-    this.transactionManager = transactionManager;
-    this.instantiator = Utils.firstNonNull(instantiator, Instantiator::usingReflection);
-    this.persistor = persistor;
-    this.submitter = Utils.firstNonNull(submitter, Submitter::withDefaultExecutor);
-    this.attemptFrequency = Utils.firstNonNull(attemptFrequency, () -> Duration.of(2, MINUTES));
-    this.blockAfterAttempts = blockAfterAttempts < 1 ? 5 : blockAfterAttempts;
-    this.flushBatchSize = flushBatchSize < 1 ? DEFAULT_FLUSH_BATCH_SIZE : flushBatchSize;
-    this.clockProvider = Utils.firstNonNull(clockProvider, () -> DefaultClockProvider.INSTANCE);
-    this.listener = Utils.firstNonNull(listener, () -> new TransactionOutboxListener() {});
-    this.logLevelTemporaryFailure = Utils.firstNonNull(logLevelTemporaryFailure, () -> Level.WARN);
-    this.validator = new Validator(this.clockProvider);
-    this.serializeMdc = serializeMdc == null || serializeMdc;
-    this.retentionThreshold = retentionThreshold == null ? Duration.ofDays(7) : retentionThreshold;
-    this.validator.validate(this);
-    this.persistor.migrate(transactionManager);
-  }
+public interface TransactionOutbox {
 
   /** @return A builder for creating a new instance of {@link TransactionOutbox}. */
-  public static TransactionOutboxBuilder builder() {
-    return new TransactionOutboxBuilder();
+  static TransactionOutboxBuilder builder() {
+    return TransactionOutboxImpl.builder();
   }
 
   /**
@@ -94,7 +26,8 @@ public class TransactionOutbox {
    *
    * <p>Returns a proxy of {@code T} which, when called, will instantly return and schedule a call
    * of the <em>real</em> method to occur after the current transaction is committed (as such a
-   * transaction needs to be active and accessible from {@link #transactionManager})
+   * transaction needs to be active and accessible from the transaction manager supplied to {@link
+   * TransactionOutboxBuilder#transactionManager(TransactionManager)}),
    *
    * <p>Usage:
    *
@@ -104,17 +37,17 @@ public class TransactionOutbox {
    * <p>This will write a record to the database using the supplied {@link Persistor} and {@link
    * Instantiator}, using the current database transaction, which will get rolled back if the rest
    * of the transaction is, and thus never processed. However, if the transaction is committed, the
-   * real method will be called immediately afterwards using the supplied {@link #submitter}. Should
-   * that fail, the call will be reattempted whenever {@link #flush()} is called, provided at least
-   * {@link #attemptFrequency} has passed since the time the task was last attempted.
+   * real method will be called immediately afterwards using the submitter supplied to {@link
+   * TransactionOutboxBuilder#submitter(Submitter)}. Should that fail, the call will be reattempted
+   * whenever {@link #flush()} is called, provided at least supplied {@link
+   * TransactionOutboxBuilder#attemptFrequency(Duration)} has passed since the time the task was
+   * last attempted.
    *
    * @param clazz The class to proxy.
    * @param <T> The type to proxy.
    * @return The proxy of {@code T}.
    */
-  public <T> T schedule(Class<T> clazz) {
-    return schedule(clazz, null);
-  }
+  <T> T schedule(Class<T> clazz);
 
   /**
    * Starts building a schedule request with parameterization. See {@link
@@ -122,21 +55,20 @@ public class TransactionOutbox {
    *
    * @return Builder.
    */
-  public ParameterizedScheduleBuilder with() {
-    return new ParameterizedScheduleBuilderImpl();
-  }
+  ParameterizedScheduleBuilder with();
 
   /**
    * Identifies any stale tasks queued using {@link #schedule(Class)} (those which were queued more
-   * than {@link #attemptFrequency} ago and have been tried less than {@link #blockAfterAttempts}
-   * times) and attempts to resubmit them.
+   * than supplied {@link TransactionOutboxBuilder#attemptFrequency(Duration)} ago and have been
+   * tried less than {@link TransactionOutboxBuilder#blockAfterAttempts(int)} )} times) and attempts
+   * to resubmit them.
    *
-   * <p>As long as {@link #submitter} is non-blocking (e.g. uses a bounded queue with a {@link
-   * java.util.concurrent.RejectedExecutionHandler} which throws such as {@link
-   * java.util.concurrent.ThreadPoolExecutor.AbortPolicy}), this method will return quickly.
-   * However, if {@link #submitter} uses a bounded queue with a blocking policy, this method could
-   * block for a long time, depending on how long the scheduled work takes and how large {@link
-   * #flushBatchSize} is.
+   * <p>As long as the {@link TransactionOutboxBuilder#submitter(Submitter)} is non-blocking (e.g.
+   * uses a bounded queue with a {@link java.util.concurrent.RejectedExecutionHandler} which throws
+   * such as {@link java.util.concurrent.ThreadPoolExecutor.AbortPolicy}), this method will return
+   * quickly. However, if the {@link TransactionOutboxBuilder#submitter(Submitter)} uses a bounded
+   * queue with a blocking policy, this method could block for a long time, depending on how long
+   * the scheduled work takes and how large {@link TransactionOutboxBuilder#flushBatchSize(int)} is.
    *
    * <p>Calls {@link TransactionManager#inTransactionReturns(TransactionalSupplier)} to start a new
    * transaction for the fetch.
@@ -147,59 +79,7 @@ public class TransactionOutbox {
    * @return true if any work was flushed.
    */
   @SuppressWarnings("UnusedReturnValue")
-  public boolean flush() {
-    Instant now = clockProvider.getClock().instant();
-    List<TransactionOutboxEntry> batch = flush(now);
-    expireIdempotencyProtection(now);
-    return !batch.isEmpty();
-  }
-
-  private List<TransactionOutboxEntry> flush(Instant now) {
-    log.debug("Flushing stale tasks");
-    var batch =
-        transactionManager.inTransactionReturns(
-            transaction -> {
-              List<TransactionOutboxEntry> result = new ArrayList<>(flushBatchSize);
-              uncheckedly(() -> persistor.selectBatch(transaction, flushBatchSize, now))
-                  .forEach(
-                      entry -> {
-                        log.debug("Reprocessing {}", entry.description());
-                        try {
-                          pushBack(transaction, entry);
-                          result.add(entry);
-                        } catch (OptimisticLockException e) {
-                          log.debug("Beaten to optimistic lock on {}", entry.description());
-                        }
-                      });
-              return result;
-            });
-    log.debug("Got batch of {}", batch.size());
-    batch.forEach(this::submitNow);
-    log.debug("Submitted batch");
-    return batch;
-  }
-
-  private void expireIdempotencyProtection(Instant now) {
-    long totalRecordsDeleted = 0;
-    int recordsDeleted;
-    do {
-      recordsDeleted =
-          transactionManager.inTransactionReturns(
-              tx ->
-                  uncheckedly(() -> persistor.deleteProcessedAndExpired(tx, flushBatchSize, now)));
-      totalRecordsDeleted += recordsDeleted;
-    } while (recordsDeleted > 0);
-    if (totalRecordsDeleted > 0) {
-      long s = retentionThreshold.toSeconds();
-      String duration = String.format("%dd:%02dh:%02dm", s / 3600, (s % 3600) / 60, (s % 60));
-      log.info(
-          "Expired idempotency protection on {} requests completed more than {} ago",
-          totalRecordsDeleted,
-          duration);
-    } else {
-      log.debug("No records found to delete as of {}", now);
-    }
-  }
+  boolean flush();
 
   /**
    * Unblocks a blocked entry and resets the attempt count so that it will be retried again.
@@ -209,19 +89,7 @@ public class TransactionOutbox {
    * @return True if the request to unblock the entry was successful. May return false if another
    *     thread unblocked the entry first.
    */
-  public boolean unblock(String entryId) {
-    if (!(transactionManager instanceof ThreadLocalContextTransactionManager)) {
-      throw new UnsupportedOperationException(
-          "This method requires a ThreadLocalContextTransactionManager");
-    }
-    log.info("Unblocking entry {} for retry.", entryId);
-    try {
-      return ((ThreadLocalContextTransactionManager) transactionManager)
-          .requireTransactionReturns(tx -> persistor.unblock(tx, entryId));
-    } catch (Exception e) {
-      throw (RuntimeException) Utils.uncheckAndThrow(e);
-    }
-  }
+  boolean unblock(String entryId);
 
   /**
    * Clears a failed entry of its failed state and resets the attempt count so that it will be
@@ -235,57 +103,7 @@ public class TransactionOutbox {
    *     thread unblocked the entry first.
    */
   @SuppressWarnings({"unchecked", "rawtypes"})
-  public boolean unblock(String entryId, Object transactionContext) {
-    if (!(transactionManager instanceof ParameterContextTransactionManager)) {
-      throw new UnsupportedOperationException(
-          "This method requires a ParameterContextTransactionManager");
-    }
-    log.info("Unblocking entry {} for retry", entryId);
-    try {
-      if (transactionContext instanceof Transaction) {
-        return persistor.unblock((Transaction) transactionContext, entryId);
-      }
-      Transaction transaction =
-          ((ParameterContextTransactionManager) transactionManager)
-              .transactionFromContext(transactionContext);
-      return persistor.unblock(transaction, entryId);
-    } catch (Exception e) {
-      throw (RuntimeException) Utils.uncheckAndThrow(e);
-    }
-  }
-
-  private <T> T schedule(Class<T> clazz, String uniqueRequestId) {
-    return Utils.createProxy(
-        clazz,
-        (method, args) ->
-            uncheckedly(
-                () -> {
-                  var extracted = transactionManager.extractTransaction(method, args);
-                  TransactionOutboxEntry entry =
-                      newEntry(
-                          extracted.getClazz(),
-                          extracted.getMethodName(),
-                          extracted.getParameters(),
-                          extracted.getArgs(),
-                          uniqueRequestId);
-                  validator.validate(entry);
-                  persistor.save(extracted.getTransaction(), entry);
-                  extracted
-                      .getTransaction()
-                      .addPostCommitHook(
-                          () -> {
-                            listener.scheduled(entry);
-                            submitNow(entry);
-                          });
-                  log.debug(
-                      "Scheduled {} for running after transaction commit", entry.description());
-                  return null;
-                }));
-  }
-
-  private void submitNow(TransactionOutboxEntry entry) {
-    submitter.submit(entry, this::processNow);
-  }
+  boolean unblock(String entryId, Object transactionContext);
 
   /**
    * Processes an entry immediately in the current thread. Intended for use in custom
@@ -294,132 +112,26 @@ public class TransactionOutbox {
    * @param entry The entry.
    */
   @SuppressWarnings("WeakerAccess")
-  public void processNow(TransactionOutboxEntry entry) {
-    try {
-      var success =
-          transactionManager.inTransactionReturnsThrows(
-              transaction -> {
-                if (!persistor.lock(transaction, entry)) {
-                  return false;
-                }
-                log.info("Processing {}", entry.description());
-                invoke(entry, transaction);
-                if (entry.getUniqueRequestId() == null) {
-                  persistor.delete(transaction, entry);
-                } else {
-                  log.debug(
-                      "Deferring deletion of {} by {}", entry.description(), retentionThreshold);
-                  entry.setProcessed(true);
-                  entry.setNextAttemptTime(after(retentionThreshold));
-                  persistor.update(transaction, entry);
-                }
-                return true;
-              });
-      if (success) {
-        log.info("Processed {}", entry.description());
-        listener.success(entry);
-      } else {
-        log.debug("Skipped task {} - may be locked or already processed", entry.getId());
-      }
-    } catch (InvocationTargetException e) {
-      updateAttemptCount(entry, e.getCause());
-    } catch (Exception e) {
-      updateAttemptCount(entry, e);
-    }
-  }
-
-  private void invoke(TransactionOutboxEntry entry, Transaction transaction)
-      throws NoSuchMethodException, IllegalAccessException, InvocationTargetException {
-    Object instance = instantiator.getInstance(entry.getInvocation().getClassName());
-    log.debug("Created instance {}", instance);
-    transactionManager.injectTransaction(entry.getInvocation(), transaction).invoke(instance);
-  }
-
-  private TransactionOutboxEntry newEntry(
-      Class<?> clazz, String methodName, Class<?>[] params, Object[] args, String uniqueRequestId) {
-    return TransactionOutboxEntry.builder()
-        .id(UUID.randomUUID().toString())
-        .invocation(
-            new Invocation(
-                instantiator.getName(clazz),
-                methodName,
-                params,
-                args,
-                serializeMdc && (MDC.getMDCAdapter() != null) ? MDC.getCopyOfContextMap() : null))
-        .nextAttemptTime(after(attemptFrequency))
-        .uniqueRequestId(uniqueRequestId)
-        .build();
-  }
-
-  private void pushBack(Transaction transaction, TransactionOutboxEntry entry)
-      throws OptimisticLockException {
-    try {
-      entry.setNextAttemptTime(after(attemptFrequency));
-      validator.validate(entry);
-      persistor.update(transaction, entry);
-    } catch (OptimisticLockException e) {
-      throw e;
-    } catch (Exception e) {
-      Utils.uncheckAndThrow(e);
-    }
-  }
-
-  private Instant after(Duration duration) {
-    return clockProvider.getClock().instant().plus(duration).truncatedTo(MILLIS);
-  }
-
-  private void updateAttemptCount(TransactionOutboxEntry entry, Throwable cause) {
-    try {
-      entry.setAttempts(entry.getAttempts() + 1);
-      var blocked = entry.getAttempts() >= blockAfterAttempts;
-      entry.setBlocked(blocked);
-      entry.setNextAttemptTime(after(attemptFrequency));
-      validator.validate(entry);
-      transactionManager.inTransactionThrows(transaction -> persistor.update(transaction, entry));
-      listener.failure(entry, cause);
-      if (blocked) {
-        log.error(
-            "Blocking failing entry {} after {} attempts: {}",
-            entry.getId(),
-            entry.getAttempts(),
-            entry.description(),
-            cause);
-        listener.blocked(entry, cause);
-      } else {
-        logAtLevel(
-            log,
-            logLevelTemporaryFailure,
-            "Temporarily failed to process entry {} : {}",
-            entry.getId(),
-            entry.description(),
-            cause);
-      }
-    } catch (Exception e) {
-      log.error(
-          "Failed to update attempt count for {}. It may be retried more times than expected.",
-          entry.description(),
-          e);
-    }
-  }
+  void processNow(TransactionOutboxEntry entry);
 
   /** Builder for {@link TransactionOutbox}. */
   @ToString
-  public static class TransactionOutboxBuilder {
+  abstract class TransactionOutboxBuilder {
 
-    private TransactionManager transactionManager;
-    private Instantiator instantiator;
-    private Submitter submitter;
-    private Duration attemptFrequency;
-    private int blockAfterAttempts;
-    private int flushBatchSize;
-    private ClockProvider clockProvider;
-    private TransactionOutboxListener listener;
-    private Persistor persistor;
-    private Level logLevelTemporaryFailure;
-    private Boolean serializeMdc;
-    private Duration retentionThreshold;
+    protected TransactionManager transactionManager;
+    protected Instantiator instantiator;
+    protected Submitter submitter;
+    protected Duration attemptFrequency;
+    protected int blockAfterAttempts;
+    protected int flushBatchSize;
+    protected ClockProvider clockProvider;
+    protected TransactionOutboxListener listener;
+    protected Persistor persistor;
+    protected Level logLevelTemporaryFailure;
+    protected Boolean serializeMdc;
+    protected Duration retentionThreshold;
 
-    TransactionOutboxBuilder() {}
+    protected TransactionOutboxBuilder() {}
 
     /**
      * @param transactionManager Provides {@link TransactionOutbox} with the ability to start,
@@ -558,24 +270,10 @@ public class TransactionOutbox {
      *
      * @return The outbox implementation.
      */
-    public TransactionOutbox build() {
-      return new TransactionOutbox(
-          transactionManager,
-          instantiator,
-          submitter,
-          attemptFrequency,
-          blockAfterAttempts,
-          flushBatchSize,
-          clockProvider,
-          listener,
-          persistor,
-          logLevelTemporaryFailure,
-          serializeMdc,
-          retentionThreshold);
-    }
+    public abstract TransactionOutbox build();
   }
 
-  public interface ParameterizedScheduleBuilder {
+  interface ParameterizedScheduleBuilder {
 
     /**
      * Specifies a unique id for the request. This defaults to {@code null}, but if non-null, will
@@ -609,23 +307,5 @@ public class TransactionOutbox {
      * @return The proxy of {@code T}.
      */
     <T> T schedule(Class<T> clazz);
-  }
-
-  private class ParameterizedScheduleBuilderImpl implements ParameterizedScheduleBuilder {
-
-    @Length(max = 250)
-    private String uniqueRequestId;
-
-    @Override
-    public ParameterizedScheduleBuilder uniqueRequestId(String uniqueRequestId) {
-      this.uniqueRequestId = uniqueRequestId;
-      return this;
-    }
-
-    @Override
-    public <T> T schedule(Class<T> clazz) {
-      validator.validate(this);
-      return TransactionOutbox.this.schedule(clazz, uniqueRequestId);
-    }
   }
 }
