+++ transactionoutbox-core/src/test/java/com/gruelbox/transactionoutbox/TestDefaultPersistorMySql8.java
@@ -1,13 +1,11 @@
 package com.gruelbox.transactionoutbox;
 
 import java.time.Duration;
-import lombok.extern.slf4j.Slf4j;
 import org.testcontainers.containers.JdbcDatabaseContainer;
 import org.testcontainers.containers.MySQLContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
-@Slf4j
 @Testcontainers
 class TestDefaultPersistorMySql8 extends AbstractDefaultPersistorTest {
