+++ transactionoutbox-core/src/test/java/com/gruelbox/transactionoutbox/acceptance/TestStubbing.java
@@ -23,12 +23,10 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import lombok.Value;
-import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 /** Checks that stubbing {@link TransactionOutbox} works cleanly. */
-@Slf4j
 class TestStubbing {
 
   @Test
